tg02/requirements.txt
aiogram==3.0.0b7
gtts==2.3.2
googletrans==4.0.0rc1
python-dotenv==1.0.0